refactor(community): modularize change password form validation logic
- Moved change password validation schema and default values to a new `change-password-form-schema.ts` file for better organization and separation of concerns.
- Updated `change-password-form.tsx` to utilize the new schema, enhancing readability and maintainability of the form component.
